+++ examples/simple_3l_speed_control.jl
@@ -1,14 +1,11 @@
-using Revise, KiteModels, OrdinaryDiffEqCore, OrdinaryDiffEqBDF, OrdinaryDiffEqSDIRK, LinearAlgebra, Timers, Statistics
+using Revise, KiteModels, LinearAlgebra
 using Base: summarysize
-tic()
 
 using Pkg
 if ! ("ControlPlots" ∈ keys(Pkg.project().dependencies))
     using TestEnv; TestEnv.activate()
 end
-using ControlPlots
-
-# TODO: sometimes very bad sim
+using ControlPlots, StatsBase
 
 set = deepcopy(load_settings("system_3l.yaml"))
 # set.elevation = 71
@@ -18,7 +15,7 @@ total_time = 10.0
 steps = Int(round(total_time / dt))
 logger = Logger(3*set.segments + 6, steps)
 
-if !@isdefined s; s = KPS4_3L(KCU(set)); end
+s::KPS4_3L = KPS4_3L(KCU(set))
 s.set = update_settings()
 s.set.abs_tol = 0.0006
 s.set.rel_tol = 0.001
@@ -33,7 +30,6 @@ sys_state = KiteModels.SysState(s)
 
 println("stepping")
 total_step_time = 0.0
-toc()
 steering = init_set_values
 amount = 0.3
 sign = 1