+++ examples_3d/parking_4p.jl
@@ -1,5 +1,7 @@
-# TODO: calculate the rotation between q and q2
-# - first add the function to calculate the rotation between two quaternions
+# Steer the kite to the side and then back until heading is about zero.
+# To run this script, use the following commands:
+# julia --project
+# include("examples_3d/parking_4p.jl")
 
 using Printf
 
@@ -7,8 +9,8 @@ using Pkg, Timers
 tic()
 if ! ("KiteViewers" ∈ keys(Pkg.project().dependencies))
     Pkg.activate("examples_3d")
-    pkg"add KiteUtils#main"
-    pkg"add KiteModels#azimuth"
+    # pkg"add KiteUtils#main"
+    # pkg"add KiteModels#azimuth"
 end
 using KiteModels, KitePodModels, KiteUtils, Rotations, StaticArrays
 using ControlPlots, KiteViewers
@@ -20,7 +22,7 @@ set = deepcopy(se())
 dt = 0.05
 set.solver="DFBDF"              # IDA or DFBDF
 set.linear_solver="GMRES"       # GMRES, LapackDense or Dense
-STEPS = 328
+STEPS = 352
 PRINT = false
 STATISTIC = false
 PLOT=false